requirements.txt
python-dotenv~=1.0.0
requests~=2.31.0
# openai==1.3.4
openai~=0.27.8
fastapi==0.104.1
uvicorn==0.24.0.post1
GitPython~=3.1.18
PyYAML==6.0.1
pydantic~=2.5.1